test(resize): cover window sizing semantics
